[refactor] playground dto auth -> playground 패키지로 이 (#341)
