+++ src/main/java/org/sopt/app/application/friend/friend_recommender/UniversityFriendRecommendRequestCreator.java
@@ -1,12 +1,12 @@
 package org.sopt.app.application.friend.friend_recommender;
 
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.OwnPlaygroundProfile;
-import org.sopt.app.application.auth.dto.RecommendFriendRequest;
+import org.sopt.app.application.playground.dto.PlaygroundProfileInfo.OwnPlaygroundProfile;
+import org.sopt.app.application.playground.dto.PlaygroundUserFindCondition;
 
 public class UniversityFriendRecommendRequestCreator implements FriendRecommendRequestCreator {
 
     @Override
-    public RecommendFriendRequest create(OwnPlaygroundProfile userPlaygroundProfile) {
+    public PlaygroundUserFindCondition createRequest(OwnPlaygroundProfile userPlaygroundProfile) {
         return null;
     }
