--- src/main/java/org/sopt/app/application/playground/PlaygroundUserRecommender.java
@@ -1,8 +1,0 @@
-package org.sopt.app.application.playground;
-
-import java.util.List;
-import org.sopt.app.application.auth.dto.RecommendFriendRequest;
-
-public interface PlaygroundUserRecommender {
-    List<Long> getPlaygroundUserIdsForSameRecommendType(final RecommendFriendRequest request);
-}
