+++ src/main/java/org/sopt/app/application/playground/dto/PlaygroundPostInfo.java
@@ -1,4 +1,4 @@
-package org.sopt.app.application.auth.dto;
+package org.sopt.app.application.playground.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
