+++ src/main/java/org/sopt/app/application/playground/dto/PlaygroundProfileInfo.java
@@ -1,4 +1,4 @@
-package org.sopt.app.application.auth.dto;
+package org.sopt.app.application.playground.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Comparator;
