+++ src/main/java/org/sopt/app/interfaces/postgres/RecommendedUserIdsRepository.java
@@ -2,11 +2,11 @@ package org.sopt.app.interfaces.postgres;
 
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
-import org.sopt.app.domain.entity.RecommendedUsers;
+import org.sopt.app.domain.entity.RecommendedUserIds;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RecommendedUserIdsRepository extends CrudRepository<RecommendedUsers, String> {
+public interface RecommendedUserIdsRepository extends CrudRepository<RecommendedUserIds, String> {
 
     @NotNull
-    Optional<RecommendedUsers> findByCondition(@NotNull String condition);
+    Optional<RecommendedUserIds> findByCondition(@NotNull String condition);
 }
