+++ src/main/java/org/sopt/app/presentation/user/UserResponseMapper.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo;
+import org.sopt.app.application.playground.dto.PlaygroundProfileInfo;
 import org.sopt.app.application.operation.OperationInfo;
 import org.sopt.app.application.app_service.AppServiceInfo;
 import org.sopt.app.application.user.UserInfo;
